add check that user id from delete request equals user id on helpPost

+++ functions/index.js
@@ -262,7 +262,12 @@ exports.reportedPostsCreate = functions.region('europe-west1').firestore.documen
     try {
       const db = admin.firestore();
       const snapValue = snap.data();
-      const { askForHelpId } = snapValue;
+      const { askForHelpId, uid:uidFromDeleteRequest } = snapValue;
+      const askForHelpCollectionEntry = await db.collection('/ask-for-help').doc(askForHelpId).get();
+      const { uid:userIdFromCollectionEntry } = askForHelpCollectionEntry;
+      if (uidFromDeleteRequest !== userIdFromCollectionEntry) {
+        return;
+      }
       await db.collection('/ask-for-help').doc(askForHelpId).delete()
     } catch (e) {
       console.error(e);
